+++ src/Toadstool/IDbCommandBuilder.cs
@@ -1,11 +1,9 @@
 using System.Collections.Generic;
 using System.Data;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace Toadstool
 {
-    public interface IDbCommandBuilder
+    public interface IDbCommandBuilder : IDbCommandExecutor
     {
         IDbCommandBuilder WithCommandText(string commandText);
         IDbCommandBuilder WithCommandTimeout(int commandTimeout);
@@ -13,8 +11,5 @@ namespace Toadstool
         IDbCommandBuilder WithParameter(string key, object value);
         IDbCommandBuilder WithParameters(object parameters);
         IDbCommandBuilder WithParameters(IDictionary<string, object> parameters);
-        Task<List<T>> ToListAsync<T>(CancellationToken cancellationToken = default(CancellationToken));
-        Task<int> ExecuteNonQueryAsync(CancellationToken cancellationToken = default(CancellationToken));
-        Task<T> ExecuteScalarAsync<T>(CancellationToken cancellationToken = default(CancellationToken));
     }
 }
