+++ src/Toadstool/IDbContext.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Data;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -7,7 +6,7 @@ namespace Toadstool
 {
     public interface IDbContext : IDisposable
     {
-        IDbCommandBuilder Query(string commandText);
+        IDbCommandBuilder Command(string commandText);
         Task<IDbTransactionWrapper> BeginTransactionAsync(CancellationToken cancellationToken = default(CancellationToken));
     }
 }
